Split agent router parity gates
